requirements.txt
opencv-python~=4.4.0.46
numpy~=1.18.5
matplotlib~=3.3.3
Pillow~=8.0.1
tensorflow~=2.3.1
fastapi
pydantic
uvicorn